Extract lsp
